backend/requirements.txt
fastapi==0.112.0
include==0.2.2
langchain==0.2.13
langchain_anthropic==0.1.22
langchain_community==0.2.12
langchain_core==0.2.30
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
modal==0.62.236
pendulum==3.0.0
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.1
varname==0.13.3
